requirements.txt
pyserial
pyvisa
pymeasure
numpy
matplotlib
minimalmodbus
pyqtgraph
pandas
multiprocess
